Hid the api key
